fixed turns life recovery formula
and made it more flexible and self explanatory (when it comes to options adjustments) at the same time. With default values it gives exactly the same outcome as mugen 1.1 (life recovery range: 12.5%-40%, works also in matches with unlimited time). With adjustments it can be turned to life recovery systems known from commercial games (example: changing TurnsRecoveryBase in options to 0 and TurnsRecoveryBonus to 20% results in life recovery range like in CvS2 ratio system: 0-20%)

sys.turnsRecoveryRate is no longer used in the formula, so I've axed it from the source.