mugen DefenceMulSet behaviour (ikemenversion)
mugen DefenceMulSet bugged behaviour (not working on first hit) is now in effect, but only if ikemenversion flag is missing. Also adjusted lifebar top layer rendering to not be changed instantly but shrink in few frames (which is also the case in mugen) in order to make popular mugen LifeAdd workaround for DefenceMulSet behaviour look more seamlessly.