Revert "Explod under parameter locked behind ikemenversion"
Unlike ModifyExplod arguments, which are often there due to author laziness (adopting existing Explod controller, just with changed name to ModifyExplod), or unawareness (it's not mentioned on Elecbyte docs which parameters are not supported), under being present in Redhots Marvel 2 characters seems like a mistake that should be handled by editing character itself. Doubt it's a widespread problem.