Fix PS3 controller detection
Joystick configuration is done by observing transitions from
"some button is pressed" to "all buttons are released" state.
See function "f_keyCfg" in "external/script/options.lua".

But right now function "getJoystickKey" returns "-11" for an
idle PS3 controller in Linux when no buttons are pressed
(because axes 2 and 5 have -1.0 value in this state). The
game does not see the "all buttons are released" state and
this breaks the joystick configuration process.

The Sony's official PS3 controller driver for Windows has
a different incompatible axes/buttons layout, but suffers
from the same problem.

This patch resolves the problem by just skipping the
problematic axes. So that the "getJoystickKey" function
returns "" in an idle state as expected.