Make all parameters in [Files] from chars .def optional
This commit improves AttachedChar portability by making all parameters in [Files] from chars .def optional. The idea is, then, to only use the st parameter to make function-only attached chars.

Recovery command has to be added inside common.cmd because, if you declare stcommon but not cmd, the game will crash in the absence of that command.