Set BGPalFX to override PalFX BGCtrl when active
This is a temporary "fix" to an issue related to PalFX BGCtrl.

How BGPalFX synthesizing should really work: when BGPalFX is applied to a stage, PalFX BGCtrl modifications should be treated like they were part of the original sprites, so a compensation should be applied to PalFX BGCtrl beforehand based on color and invertall values.