Fixed transition between hit and idle states
Mugen has long had an issue where characters who have just exited a hit state cannot block in the same frame, appearing to be in the idle state but actually being unable to act. Hitting them in that frame (correctly) counts as a combo, even.
These changes attempt to keep the gameplay intact, but now keeping the character in a get hit animation while they are unable to move, so that the screen reflects what is truly happening.
This inadvertently fixes issue #597 because the character no longer falsely goes through an idle state.