Implement SFF caching
The cache keeps a copy of all SFF objects (not the actual sprite data) so that
they can be immediately reused when the same file is loaded again. A reference
count is kept so that the last GC’ed version of an SFF will cause the cached
copy to be deleted, too.