RatioMatches Extra Code
I am still crashing on startup without altering this syntax to how it is.
Here's the steps any layperson, I.E. me, can perform to reproduce my crash.
1: Fresh bin folder (from a new compilation of the latest IKEMEN) with no modifications. This folder has only Data, External and the latest Ikemen GO exe compiled from github.
2: Then add from the Elecbyte pack repository, the necessary System.def components.
3: Add fonts in from the pack and get my error. Voila!

Speaking with Lazin3ss on it, all I could really gather is that something's up with the syntax for rmax and rmin is written. After some troubleshooting, this got me back into the program.