Make 'command' trigger work like selfcommand only if ikemenver > 0
Since it doesn't make too much sense to have the same function working differently depending on the name, now 'command' trigger works like selfcommand, but only if the character has Ikemenversion > 0, otherwise it will work like mugen.

The reason for this is that command names are generally only known for the character itself, one character can't really tell what commands the enemy have, so it makes sense that, when cheking enemy command execution by redirecting 'command' trigger, we check if the enemy is performing one of our commands.