Rewrite explod code to improve MUGEN 1.1 compatibility
This rewrite doesn't reproduce the facing bug, and it doesn't fully reproduce the space = screen and postype = p1/p2 bug when spawning explods in a zoomed out stage, but it's closer than before.

This rewrite is kinda workaround-heavy, so a refactor is necessary to expand explod code in the future, but it works well enough for now.