feat: Resizable window, scale game to fit window
This change re-enables the ability to resize the game window and
allows Ikemen GO to scale the game's viewport to fit the window
by using a new function, GetScaledViewportSize(), and then simply updating
the viewport in EndFrame().

It's a basic scaling algorithim that just scales the GameWidth/GameHeight
to the size of the window while maintaing aspect ratio, and then applies
this to the viewport with a gl.Viewport command. Sadly the result doesn't
apply any filter when scaling, so you get nearest-neighbor scaled pixels,
but it's at least better than no resizing the window at all IMO.

Window size is not saved, and GameWidth/GameHeight still determine the
starting size of the game's resolution and window size.

I hope to give this more flexibility in the future, such as
allowing the end-user to toggle the aspect ratio correction.