feat: KeepAspect (MUGEN 1.1 feature), Y-centering to window scaling
As per K4thos's request, this adds the KeepAspect setting from MUGEN 1.1.
Defaults to "true", and when set to "false" it stretches the game to
fill the entire window, whatever size it may be.

Quite a few people also considered the lack of Y centering when resizing
the window to be a bug, so I've fixed that in this PR as well.