feat: MSAA Samples
Suggested by worleydl, this allows the user to adjust how many
samples are used by MSAA. Reducing the sample count seems to allow for
certain systems that normally render black screens with MSAA to use it.
Also added a check to make sure that the value is never 0 or lower, cause
that causes the framebuffer to not be created.

Defaults to 2, which seems to be a sane default value. Todo: add this and
my other new options to Video Settings.