refactor: Use post-process shaders *after* resizing frame
Turns out I was wrong with a commit a few weeks back and standard
order is to post-process after scaling the frame, not before.

I also set texture size to the size of the viewport rather than the
actual texture size since it appeared to fix certain shaders that would
otherwise break under the previous refactor, such as AdvancedCRT. I
could always revert that part if it turns out to be desirable, or even
expose viewport and texture as separate uniforms if that's desirable.